SigmaBackend/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.5
django-js-asset==2.2.0
django-richtextfield==1.6.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-decouple==3.8
sqlparse==0.5.0
whitenoise==6.7.0
